feat(task-graph): ship InMemoryTaskOutputRepository from ./test
The class was defined only in packages/test/src/binding, so 12 tests
could not leave that package without breaking their import. It now lives
in task-graph, which owns the base class it extends, and ships from
@workglow/task-graph/test.

This is the simpler of the two ./test shapes, and the distinction is
worth stating: a symbol the main bundle also touches (util's reset
hooks, ai's clearCheckpoints) has to route through an _internal bag so
both entries resolve to one instance. A test double is purely additive
-- nothing in the main bundle references it -- so it can be defined in
the test bundle outright, with no _internal.

What still applies to both shapes is the base-class import. Reaching
TaskOutputTabularRepository relatively would inline a second copy into
this bundle and instances would fail instanceof against the class every
consumer holds -- a dist-only failure that presents as a type bug rather
than an unresolved import. The bare specifier keeps it external; the
built entry is 589 bytes with both imports external. A guard test
asserts the instanceof across the boundary.

The testing/ folder is safe from the public API because task-graph's
barrel star-exports per file, never per directory.

Build 84/84; 29 files / 165 tests pass across the dependents.
