+++ packages/task-graph/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "include": ["src/**/*"],
-  "files": ["./src/browser.ts", "./src/node.ts"],
+  "files": ["./src/browser.ts", "./src/node.ts", "./src/test-entry.ts"],
   "exclude": ["dist", "node_modules", "src/**/__tests__/**"],
   "compilerOptions": {
     "composite": true,
