+++ packages/test/src/test/task-graph-output-cache/StreamingCache.test.ts
@@ -6,11 +6,11 @@
 
 import type { CachePolicy, IExecuteContext, StreamEvent } from "@workglow/task-graph";
 import { Task, TaskStatus } from "@workglow/task-graph";
+import { InMemoryTaskOutputRepository } from "@workglow/task-graph/test";
 import { setLogger } from "@workglow/util";
 import type { DataPortSchema } from "@workglow/util/schema";
 import { getTestingLogger } from "@workglow/util/test";
 import { beforeEach, describe, expect, it } from "vitest";
-import { InMemoryTaskOutputRepository } from "../../binding/InMemoryTaskOutputRepository";
 
 // ============================================================================
 // Test Tasks
